Break out error handlers into a small utility object
